requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.9
gunicorn==19.7.1
idna==2.5
lxml==3.8.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
requests==2.18.3
Unidecode==0.4.21
urllib3==1.22
vk-api==8.8
validate_email
phonenumbers
telethon
django-celery-beat
django-celery-results
celery==4.1.0
redis
django-ipware